+++ code/lib/utility.py
@@ -7,7 +7,7 @@ from typing import Any
 def shift(img: NDArray[Any], offset: tuple[int, int]) -> NDArray[Any]:
     translation_matrix = np.array(
         [[1, 0, offset[0]], [0, 1, offset[1]]],
-        dtype=np.float32
+        dtype=np.float64
     )
     img_shift = cv.warpAffine(img.astype(np.uint8), translation_matrix, (img.shape[1],img.shape[0]))
     return img_shift
